revising

+++ Programs/4Faceprep/all/02.cpp
@@ -7,7 +7,7 @@ int main()
 
      char c;
      cin>>c;
-     if((int)c>=65 && (int)c<=90 || (int)c>=97 && (int)c<=122)
+     if(c>=65 && c<=90 || c>=97 && c<=122)
      {
          cout<<"its an alphabet\n";
      }
